refactor(examples): extract the engine sequence so errors stay short-lived
govet:shadow and gocritic:sloppyReassign are mutually exclusive for the
shape these examples had. Verified both ways at usertask-sla/main.go:51:
`if err := th.Run(ctx)` reports "shadows declaration at line 33", and
`if err = th.Run(ctx)` reports "can be replaced with err :=". No form of
that line satisfies both checks.

The cause is lifetime, not syntax. A single err threaded through
build → register → run → wait is revived by each later step, so every
scoped `if err := …` in between shadows it. Extracting the engine
sequence into runProcess gives each step's error one short scope, and
both linters go quiet as a consequence of the structure rather than by
excluding either check.

Two of the five were single-file examples with run() functions of 149
and 121 lines — process-data at 269 lines total, parallel-gateway at
192 — well past the point where an example is scannable. They split by
concern, the way the house rule asks: main.go wires, process.go builds
the model, handlers.go holds the Go operations, engine.go runs and
verifies. process-data's builder returns a small struct rather than four
values, since the run reads results back from two DataObjects.

versioning needed a different move: its handle assignment `h, err =
engine.StartLatest(…)` was what revived the outer err, so the handle got
its own error variable and the four scenario checks became scoped.

Context lifetimes are preserved rather than moved — runProcess takes ctx
as a parameter where the caller owns the cancel, so no defer changes the
order it fires in.

Verified per example and in bulk: shadow 0, sloppyReassign 0, gofmt
clean, and all 49 examples still run end to end. The linter cannot see
a defer that fires too early; running them can.
