+++ examples/parallel-gateway/handlers.go
@@ -1,0 +1,55 @@
+package main
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/dr-dobermann/gobpm/pkg/model/activities"
+	"github.com/dr-dobermann/gobpm/pkg/model/data"
+	"github.com/dr-dobermann/gobpm/pkg/model/flow"
+	"github.com/dr-dobermann/gobpm/pkg/model/service"
+	"github.com/dr-dobermann/gobpm/pkg/model/service/gooper"
+)
+
+// newWorker builds a ServiceTask whose operation is a Go functor: it prints its
+// execution and signals done, so the example shows real per-branch work rather
+// than a silent no-op.
+func newWorker(name string, done chan<- string) (*activities.ServiceTask, error) {
+	op, err := gooper.New(
+		name+"-op",
+		func(_ context.Context, _ service.DataReader, _ *data.ItemDefinition) (*data.ItemDefinition, error) {
+			fmt.Printf("  ▶ %s executed\n", name)
+			done <- name
+
+			return nil, nil
+		})
+	if err != nil {
+		return nil, fmt.Errorf("create %s operation: %w", name, err)
+	}
+
+	st, err := activities.NewServiceTask(name, op, activities.WithoutParams())
+	if err != nil {
+		return nil, fmt.Errorf("create %s task: %w", name, err)
+	}
+
+	return st, nil
+}
+
+// link connects two flow elements with a sequence flow.
+func link(src, trg flow.Element) error {
+	s, ok := src.(flow.SequenceSource)
+	if !ok {
+		return fmt.Errorf("%q is not a sequence source", src.Name())
+	}
+
+	t, ok := trg.(flow.SequenceTarget)
+	if !ok {
+		return fmt.Errorf("%q is not a sequence target", trg.Name())
+	}
+
+	if _, err := flow.Link(s, t); err != nil {
+		return fmt.Errorf("link: %w", err)
+	}
+
+	return nil
+}
